+++ dhis-2/dhis-support/dhis-support-system/src/main/java/org/hisp/dhis/system/objectmapper/DeflatedDataValueNameMinMaxRowMapper.java
@@ -1,7 +1,7 @@
 package org.hisp.dhis.system.objectmapper;
 
 /*
- * Copyright (c) 2004-2019, University of Oslo
+ * Copyright (c) 2004-2020, University of Oslo
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -83,11 +83,11 @@ public class DeflatedDataValueNameMinMaxRowMapper
     {
         final DeflatedDataValue value = new DeflatedDataValue();
         
-        value.setDataElementId( resultSet.getInt( "dataelementid" ) );
-        value.setPeriodId( resultSet.getInt( "periodid" ) );
-        value.setSourceId( resultSet.getInt( "sourceid" ) );
-        value.setCategoryOptionComboId( resultSet.getInt( "categoryoptioncomboid" ) );
-        value.setAttributeOptionComboId( resultSet.getInt( "attributeoptioncomboid" ) );
+        value.setDataElementId( resultSet.getLong( "dataelementid" ) );
+        value.setPeriodId( resultSet.getLong( "periodid" ) );
+        value.setSourceId( resultSet.getLong( "sourceid" ) );
+        value.setCategoryOptionComboId( resultSet.getLong( "categoryoptioncomboid" ) );
+        value.setAttributeOptionComboId( resultSet.getLong( "attributeoptioncomboid" ) );
         value.setValue( resultSet.getString( "value" ) );
         value.setStoredBy( resultSet.getString( "storedby" ) );
         value.setCreated( resultSet.getDate( "created" ) );
